LTP: fix testcase fchown04 issue
The original test case performs the below tests:
1) fchown(2) returns -1 and sets errno to EPERM if the effective
   user id of the process does not match the owner of the file
   and the process is not a superuser.
2) fchown(2) returns -1 and sets errno to EBADF if the file
   descriptor of the specified file is not valid.
3) fchown(2) returns -1 and sets errno to EROFS if the named file
   resides on a read-only file system.

Problem/issue:
This test case causing oom killer to be invoked and causing
kernel panic. This is because the test case invokes test framework
function "tst_acquire_device" which creates a 256MB of .img file
and binds with a loop device.  This is not supported because the
default size of LKL memory is set to 32M. This is kept low due to
the limited size of EPC (Enclave page cache) and to avoid page
cache being swapped out.
Also, the test case invokes "tst_mkfs" framework function which
inturn invokes mkfs utility command with the help of system() syscall.
It is recommended to use root file system, but, this test case
needs a read-only filesystem to test one of the sub-test case.
we don't have a read-only file system mounted in sgx-lkl.
Hence, disabling the code related to the mount and sub test case.

Modifications:
 1. Disable the 3rd sub test case
 2. Disable the code which creates file system, mount, and unmount
    the filesystem.