LTP: fix test case mount04 issue
Problem/issue:
This test case causing oom killer to be invoked and causing
kernel panic. This is because the test case invokes test framework
function "tst_acquire_device" which creates a 256MB of .img file
and binds with a loop device. This is not supported because the
default size of LKL memory is set to 32M. This is kept low due to
the limited size of EPC (Enclave page cache) and to avoid page
cache being swapped out.
Also, the test case invokes "tst_mkfs" framework function which
inturn invokes mkfs utility command with the help of system() syscall.
It is recommended to use root file system, but, this test case
needs a read-only filesystem to test one of the sub-test case.
we don't have a read-only file system mounted in sgx-lkl.
Hence, disabling the code related to the mount and sub test case.

Modifications:
1. Added the code to create a loop device and mount.
2. This need some changes to lsds/sgx-lkl repo, new PR is raised lsds/sgx-lkl#741